requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy>=1.22.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.64.0